pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask==3.0.0",
    "flask-sqlalchemy==3.1.1",
    "Flask-Mail==0.9.1",
    "Werkzeug==3.0.1",
    "reportlab==4.0.4",
    "gunicorn==21.2.0",
    "flask-talisman==1.1.0",
    "bleach==6.1.0",
    "requests==2.31.0",
    "markupsafe>=3.0.2",
    "sqlalchemy>=2.0.41",
    "psycopg2-binary>=2.9.10",
    "email-validator>=2.2.0",
    "anthropic>=0.58.2",
    "openai>=1.97.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "replit-object-storage>=1.0.2",
    "pillow>=11.3.0",
    "flask-wtf>=1.2.2",
    "flask-limiter>=3.12",
    "fitz>=0.0.1.dev2",
    "pymupdf>=1.26.4",
]
